microservices/assessment-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.1
drf-spectacular==0.26.5
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
python-decouple==3.8
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
